refactor(lib): add `IncomingChannel` and `OutgoingChannel`, change signature of event listeners, add `Event`, etc.

+++ onebot-11-kotlin-lib/src/main/kotlin/cn/chuanwise/onebot/lib/v11/data/event/NoticeEventData.kt
@@ -162,7 +162,7 @@ data class GroupMessageRecallEventData(
     val groupID: Long,
     val userID: Long,
     val operatorID: Long,
-    val messageID: Long,
+    val messageID: Int,
 ) : NoticeEventData
 
 
@@ -177,7 +177,7 @@ data class FriendMessageRecallEventData(
     override val noticeType: String,
 
     val userID: Long,
-    val messageID: Long,
+    val messageID: Int,
 ) : NoticeEventData
